move constant to CCDConfigConstant
