+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/documentcatergory/DraftOrdersCategoriser.java
@@ -14,6 +14,7 @@ import java.util.Optional;
 
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.UPLOAD_PARTY_APPLICANT;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.UPLOAD_PARTY_RESPONDENT;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.SUGGESTED_DRAFT_ORDER_OPTION;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocumentParty.APPLICANT;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseDocumentParty.RESPONDENT;
 
@@ -48,7 +49,7 @@ public class DraftOrdersCategoriser {
     }
 
     private boolean isSuggestedDraftOrderPriorToHearing(FinremCaseData finremCaseData) {
-        return "aSuggestedDraftOrderPriorToAListedHearing".equals(finremCaseData.getDraftOrdersWrapper().getTypeOfDraftOrder());
+        return SUGGESTED_DRAFT_ORDER_OPTION.equals(finremCaseData.getDraftOrdersWrapper().getTypeOfDraftOrder());
     }
 
     private String determineChosenParty(FinremCaseData finremCaseData, String userRole) {
